first implementation of HLS DOS
created a HlsDos.cs which uses the AdfFile.cs class to read the relevant sectors of an adf file. HLS is the Amiga "Hockey Leage Simulator" for the "Wayne Gretzky Hockey" games. Unfortunately the developers created an own DOS format for their save discs.
This first implementation will be able to read the directory of the save disc.